QOL Changes Pawns Can now move 2 squares on their first turn Turns

+++ src/BoardPiece.java
@@ -6,6 +6,7 @@ public class BoardPiece {
     public boolean isWhite;
     public LinkedList<BoardPiece> pieces;
     public int x, y;
+    public boolean hasMoved;
 
     public BoardPiece(int xPos, int yPos, boolean isWhite, LinkedList<BoardPiece> pieces, int scale) {
         this.xPos = xPos;
@@ -16,10 +17,18 @@ public class BoardPiece {
         this.isWhite = isWhite;
         this.pieces = pieces;
         pieces.add(this);
+        hasMoved = false;
     }
 
     public void move(int xPos, int yPos) {
         BoardPiece piece = Main.getPiece(xPos * scale, yPos * scale);
+        LinkedList<Vector2d> legalSquares = getLegalSquares();
+        Vector2d testVector = new Vector2d(xPos, yPos);
+        if (!legalSquares.contains(testVector) || !(Main.isWhiteTurn == isWhite)) {
+            x = this.xPos * scale;
+            y = this.yPos * scale;
+            return;
+        }
         if (piece != null) {
             if (piece.isWhite != isWhite) {
                 piece.kill();
@@ -33,6 +42,8 @@ public class BoardPiece {
         this.yPos = yPos;
         x = xPos * scale;
         y = yPos * scale;
+        if (!hasMoved) hasMoved = true;
+        Main.isWhiteTurn = !Main.isWhiteTurn;
     }
 
     public void kill() {
@@ -41,10 +52,12 @@ public class BoardPiece {
 
     public LinkedList<Vector2d> getLegalSquares() {
         return null;
-    };
+    }
 
     public void printMoves(LinkedList<Vector2d> moves) {
-
+        for (Vector2d move : moves) {
+            System.out.print(move.x + " " + move.y + " ");
+        }
     }
 
     public BoardPiece getPieceVec2D(Vector2d vec) {;
